Implement default on kdf
