+++ crates/bitwarden-crypto/src/keys/mod.rs
@@ -1,7 +1,10 @@
 mod key_encryptable;
 pub use key_encryptable::{CryptoKey, KeyDecryptable, KeyEncryptable};
 mod master_key;
-pub use master_key::{HashPurpose, Kdf, MasterKey};
+pub use master_key::{
+    default_argon2_iterations, default_argon2_memory, default_argon2_parallelism,
+    default_pbkdf2_iterations, HashPurpose, Kdf, MasterKey,
+};
 mod shareable_key;
 pub use shareable_key::derive_shareable_key;
 mod symmetric_crypto_key;
