+++ crates/bitwarden/src/auth/login/auth_request.rs
@@ -2,6 +2,7 @@ use bitwarden_api_api::{
     apis::auth_requests_api::{auth_requests_id_response_get, auth_requests_post},
     models::{AuthRequestCreateRequestModel, AuthRequestType},
 };
+use bitwarden_crypto::Kdf;
 use uuid::Uuid;
 
 use crate::{
@@ -12,7 +13,6 @@ use crate::{
     client::{LoginMethod, UserLoginMethod},
     error::{require, Result},
     mobile::crypto::{AuthRequestMethod, InitUserCryptoMethod, InitUserCryptoRequest},
-    util::default_kdf,
     Client,
 };
 
@@ -84,7 +84,7 @@ pub(crate) async fn complete_auth_request(
     .await?;
 
     if let IdentityTokenResponse::Authenticated(r) = response {
-        let kdf = default_kdf();
+        let kdf = Kdf::default();
 
         client.set_tokens(
             r.access_token.clone(),
